feat: Add children prop to PropsOptions interface
This commit adds the  prop to the  interface in the  file. The  prop is a function that takes a boolean parameter  and returns a JSX element. This change allows for more flexible rendering of components that utilize the  interface.

Note: This commit message follows the conventional commit format, starting with a type (feat) followed by a brief description of the changes.